refactor: consolidate BaseController hierarchy into single unified class
This change merges three separate base controller classes into one:
- BaseController (basic error handling, pagination, filters)
- EnhancedBaseController (logging, request context, operational errors)
- AuditedBaseController (audit context injection)

The new unified BaseController:
- Uses composable options (enableAudit, enableLogging) instead of inheritance
- Supports both legacy and enhanced method signatures for backward compatibility
- Exports RequestWithContext, OperationalError, and ControllerOptions types
- Provides lazy-initialized audit database access

Migration summary:
- 5 controllers migrated from EnhancedBaseController with enableLogging: true
- 2 controllers migrated from AuditedBaseController with enableAudit: true
- Remaining controllers use BaseController with default options

Fixes #30
